Fix deprecation

+++ src/main/kotlin/at/bitfire/dav4jvm/ktor/exception/HttpResponseInfo.kt
@@ -83,7 +83,7 @@ internal class HttpResponseInfo private constructor(
                     try {
                         val responseBody = responseBodyChannel ?: response.bodyAsChannel()
                         val charset = responseContentType.charset() ?: Charsets.UTF_8
-                        responseBody.readBuffer(MAX_EXCERPT_SIZE).readString(charset)
+                        responseBody.readBuffer(MAX_EXCERPT_SIZE.toLong()).readString(charset)
                     } catch (_: Exception) {
                         // response body not available anymore, probably already consumed
                         null
